Reorganize space game packages
Changes
- Split space game scripts into movement and combat packages
- Move mayonez.scripts.combat classes to spacegame.combat package
